Add and remove and clear all feature
