+++ src/index.html
@@ -25,7 +25,7 @@
     <ul class="task-list">
      
     </ul>
-    <button class="clear-button">Clear all task</button>
+    <button type="button" class="clear-button">Clear all task</button>
 </div>
 
 </html>
