+++ src/index.js
@@ -9,9 +9,9 @@ import './style.css';
 import Todo from './Todo.js';
 
 // Events: display Task
-document.addEventListener('DOMContentLoaded', Store.displayTask);
-
+// document.addEventListener('DOMContentLoaded', Store.displayTask);
 // Events: add a Task
+
 form.addEventListener('submit', (e) => {
   e.preventDefault();
   const index = Store.getTask().length + 1;
